Refactor preloader to dottedPreloader and add abstract like class Preloader
