+++ src/dataLoader.js
@@ -1,7 +1,7 @@
-const dataLoader = async (loader, requester) => {
-    loader.render();
+const dataLoader = async (preloader, requester) => {
+    preloader.show();
     const data = await requester.getData();
-    loader.clear();
+    preloader.hide();
     return data;
 }
